From Tom Jolly.  a couple of small changes to osgParticle.
"The first is with
Particle.cpp.  I made a change so that when the lifetime is 0 (lasts
forever) the sizeRange, colorRange, and alphaRange are used to create
a random size, color, and alpha."

"The second change is with range and SectorPlacer.  The distribution of
particles across the sector was not uniform.  I added get_random_sqrtf()
function where it is used in SectorPlacer::Place().  This seems to make
the distribution uniform (at least when minimum radius is 0)."