From Charles Cole, "he attached code implements the LightPointSystem class to allow for the OpenFlight plug-in to read and handle light point system nodes.  The behavior is very similar to the old plug-in in that a MultiSwitch node is created to handle the "enabled" flag bit set in the node record.  The code also reverts the changes for the actualPixelSize as mentioned above.  And lastly, the code requires the previously submitted changes for the plug-in.
As for the other changes, I've tested the code with Visual Studio 2005
and the files that I posted in the users forum.

With all of the submitted changes, the OpenFlight plug-in should now be
capable of loading files with light point system nodes and the use of
palletized light points and non-palletized light points.
"